Rename migration to 3.6.0-beta.1 and move writeEmbeddedKubeConfig to app-utils
