+++ src/common/cluster-store.ts
@@ -1,7 +1,4 @@
-import { app, remote } from "electron"
 import ElectronStore from "electron-store"
-import  { ensureDirSync, writeFileSync } from "fs-extra"
-import * as path from "path"
 import { Cluster, ClusterBaseInfo } from "../main/cluster";
 import * as version200Beta2 from "../migrations/cluster-store/2.0.0-beta.2"
 import * as version241 from "../migrations/cluster-store/2.4.1"
@@ -9,7 +6,7 @@ import * as version260Beta2 from "../migrations/cluster-store/2.6.0-beta.2"
 import * as version260Beta3 from "../migrations/cluster-store/2.6.0-beta.3"
 import * as version270Beta0 from "../migrations/cluster-store/2.7.0-beta.0"
 import * as version270Beta1 from "../migrations/cluster-store/2.7.0-beta.1"
-import * as version350Beta1 from "../migrations/cluster-store/3.5.0-beta.1"
+import * as version360Beta1 from "../migrations/cluster-store/3.6.0-beta.1"
 import { getAppVersion } from "./utils/app-version";
 
 export class ClusterStore {
@@ -30,7 +27,7 @@ export class ClusterStore {
         "2.6.0-beta.3": version260Beta3.migration,
         "2.7.0-beta.0": version270Beta0.migration,
         "2.7.0-beta.1": version270Beta1.migration,
-        "3.5.0-beta.1": version350Beta1.migration
+        "3.6.0-beta.1": version360Beta1.migration
       }
     })
   }
@@ -121,16 +118,3 @@ export class ClusterStore {
 }
 
 export const clusterStore: ClusterStore = ClusterStore.getInstance();
-
-// Writes kubeconfigs to "embedded" store, i.e. .../Lens/kubeconfigs/
-export function writeEmbeddedKubeConfig(clusterId: string, kubeConfig: string): string {
-  // This can be called from main & renderer
-  const a = (app || remote.app)
-  const kubeConfigBase = path.join(a.getPath("userData"), "kubeconfigs")
-  ensureDirSync(kubeConfigBase)
-
-  const kubeConfigFile = path.join(kubeConfigBase, clusterId)
-  writeFileSync(kubeConfigFile, kubeConfig)
-
-  return kubeConfigFile
-}
